Assert the from_rep atom/literal range
Atom::from_rep and Literal::from_rep reconstruct a value from its raw
integer representation. Add a debug assertion that the representation is
in range -- an atom no larger than the maximum literal, a literal no
smaller than the minimum -- so an out-of-range value is caught as it
enters the type rather than where a later sign flip overflows it.

Also drop three narrowing casts left over from the crossing conversion:
the heuristic priority, the external-atom test, and the clause-literal
wrap each take their argument without one.

+++ lib/output/src/backend.cc
@@ -1588,7 +1588,7 @@ class OutputBackend : public OutputStm, OutputTheory {
                         auto elit = bld_.clause(std::array{hlit, clit}, ClauseType::conjunctive);
                         bd_conds.back().emplace_back(elit);
                         // elit is a fresh positive clause literal
-                        bld_.mark(Literal::from_rep(uid_to_lit(static_cast<size_t>(elit))), EQType::implication);
+                        bld_.mark(Literal::from_rep(uid_to_lit(elit)), EQType::implication);
                         bld_.mark(Literal::from_rep(clit), EQType::implication);
                         rule_.add(bld_, std::array{hlit}, std::array{blit, clit}, true);
                     } else {
